requirements.txt
Django==5.2.6
djangorestframework==3.16.1
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.8.0
python-dotenv==1.0.0